% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bestToOmit}
\alias{bestToOmit}
\title{Identify the columns with most missing data}
\usage{
bestToOmit(grp, omit, ref = NULL)
}
\arguments{
\item{grp}{an IFA group}

\item{omit}{the maximum number of items to omit}

\item{ref}{the reference column (optional)}
}
\description{
If a reference column is given then only rows that are not missing
on the reference column are considered. Otherwise all rows are
considered.
}

