% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{compressDataFrame}
\alias{compressDataFrame}
\title{Compress a data frame into unique rows and frequencies}
\usage{
compressDataFrame(tabdata, freqColName = "freq", .asNumeric = FALSE)
}
\arguments{
\item{tabdata}{An object of class \code{data.frame}}

\item{freqColName}{Column name to contain the frequencies}

\item{.asNumeric}{logical. Whether to cast the frequencies to the numeric type}
}
\value{
Returns a compressed data frame
}
\description{
Compress a data frame into unique rows and frequency counts.
}
\examples{
df <- as.data.frame(matrix(c(sample.int(2, 30, replace=TRUE)), 10, 3))
compressDataFrame(df)
}
