% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{crosstabTest}
\alias{crosstabTest}
\title{Monte-Carlo test for cross-tabulation tables}
\usage{
crosstabTest(ob, ex, trials)
}
\arguments{
\item{ob}{observed table}

\item{ex}{expected table}

\item{trials}{number of Monte-Carlo trials}
}
\description{
\lifecycle{experimental}
This is for developers.
}
