% Generated by roxygen2 (4.0.1): do not edit by hand
\name{EAPscores}
\alias{EAPscores}
\title{Compute EAP scores}
\usage{
EAPscores(grp, ..., naAction = "fail")
}
\arguments{
\item{grp}{a list with spec, param, and data}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{naAction}{action for rows with fewer than
\code{minItemsPerScore}. Defaults to 'fail'. If 'pass', will fill
with NAs.}
}
\description{
Compute EAP scores
}
\examples{
spec <- list()
spec[1:3] <- rpf.grm(outcomes=3)
param <- sapply(spec, rpf.rparam)
data <- rpf.sample(5, spec, param)
colnames(param) <- colnames(data)
grp <- list(spec=spec, param=param, data=data, minItemsPerScore=1L)
EAPscores(grp)
}

