% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{expandDataFrame}
\alias{expandDataFrame}
\title{Expand summary table of patterns and frequencies}
\usage{
expandDataFrame(tabdata, freqName = NULL)
}
\arguments{
\item{tabdata}{An object of class \code{data.frame} with the unique response patterns and the number of frequencies}

\item{freqName}{Column name containing the frequencies}
}
\value{
Returns a data frame with all the response patterns
}
\description{
Expand a summary table of unique response patterns to a full sized
data-set.
}
\examples{
data(LSAT7)
expandDataFrame(LSAT7, freqName="freq")
}
\author{
Based on code by Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
