% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{fromFactorLoading}
\alias{fromFactorLoading}
\title{Convert factor loadings to response function slopes}
\usage{
fromFactorLoading(loading, ogive = rpf.ogive)
}
\arguments{
\item{loading}{a matrix with items in the rows and factors in the columns}

\item{ogive}{the ogive constant (default rpf.ogive)}
}
\value{
a slope matrix with items in the columns and factors in the rows
}
\description{
Convert factor loadings to response function slopes
}

