% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fromFactorThreshold}
\alias{fromFactorThreshold}
\title{Convert factor thresholds to response function intercepts}
\usage{
fromFactorThreshold(threshold, loading, ogive = rpf.ogive)
}
\arguments{
\item{threshold}{a matrix with items in the columns and thresholds in the rows}

\item{loading}{a matrix with items in the rows and factors in the columns}

\item{ogive}{the ogive constant (default \link{rpf.ogive})}
}
\value{
an item intercept matrix with items in the columns and intercepts in the rows
}
\description{
Convert factor thresholds to response function intercepts
}
\seealso{
Other factor model equivalence: 
\code{\link{fromFactorLoading}()},
\code{\link{toFactorLoading}()},
\code{\link{toFactorThreshold}()}
}
\concept{factor model equivalence}
