% Generated by roxygen2 (4.0.2): do not edit by hand
\name{itemOutcomeBySumScore}
\alias{itemOutcomeBySumScore}
\title{Produce an item outcome by observed sum-score table}
\usage{
itemOutcomeBySumScore(grp, mask, interest)
}
\arguments{
\item{grp}{a list with spec, param, and data}

\item{mask}{a vector of logicals indicating which items to include}

\item{interest}{index or name of the item of interest}
}
\description{
Produce an item outcome by observed sum-score table
}
\examples{
set.seed(1)
spec <- list()
spec[1:3] <- rpf.grm(outcomes=3)
param <- sapply(spec, rpf.rparam)
data <- rpf.sample(5, spec, param)
colnames(param) <- colnames(data)
grp <- list(spec=spec, param=param, data=data)
itemOutcomeBySumScore(grp, c(FALSE,TRUE,TRUE), 1L)
}

