% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kct.R
\docType{data}
\name{kct}
\alias{kct}
\alias{kct.items}
\alias{kct.people}
\title{Knox Cube Test dataset}
\description{
These data from Wright & Stone (1979, p. 31) were fit with
Winsteps 3.73 using a 1PL model (slope fixed to 1).
}
\examples{
data(kct)
}
\references{
Wright, B. D. & Stone, M. H. (1979). \emph{Best Test Design: Rasch Measurement.} Univ of Chicago Social Research.
}
\keyword{data}
