% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAT.R
\docType{data}
\name{LSAT6}
\alias{LSAT6}
\title{Description of LSAT6 data}
\description{
Data from Thissen (1982); contains 5 dichotomously scored
items obtained from the Law School Admissions Test, section 6.
}
\examples{
data(LSAT6)
}
\references{
Thissen, D. (1982). Marginal maximum likelihood estimation for the one-parameter logistic model.
\emph{Psychometrika, 47}, 175-186.
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{data}
