\docType{data}
\name{mentalsilence}
\alias{mentalsilence}
\alias{ms.items}
\alias{ms.people}
\title{Mental silence dataset}
\source{
  \url{http://openscienceframework.org/project/fpxmZ/}
}
\description{
  These data are from Pritikin & Schmidt (2013, March).
}
\details{
  All items were fit to a 5 category Generalized Partial
  Credit Model (GPCM) with standard Gauss-Hermite
  quadrature of 30 points and a lognormal prior for the
  discrimination parameters using Parscale 4.1.2328.4, -2
  log likelihood = 20816.14.
}
\examples{
data(ms.items)
data(ms.people)
}
\references{
  Pritikin, J. N., & Schmidt, K. (2013, March). A
  self-report measure for familiarity with mental silence.
  Accepted for Spirituality in the 21st Century: At the
  Interface of Theory, Praxis and Pedagogy, Lisbon,
  Portugal.
}
\keyword{datasets}

