% Generated by roxygen2 (4.0.1): do not edit by hand
\name{observedSumScore}
\alias{observedSumScore}
\title{Compute the observed sum-score}
\usage{
observedSumScore(grp, mask)
}
\arguments{
\item{grp}{a list with spec, param, and data}

\item{mask}{a vector of logicals indicating which items to include}
}
\description{
Compute the observed sum-score
}
\examples{
spec <- list()
spec[1:3] <- rpf.grm(outcomes=3)
param <- sapply(spec, rpf.rparam)
data <- rpf.sample(5, spec, param)
colnames(param) <- colnames(data)
grp <- list(spec=spec, param=param, data=data)
observedSumScore(grp, rep(TRUE, length(spec)))
}

