% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{omitItems}
\alias{omitItems}
\title{Omit the given items}
\usage{
omitItems(grp, excol)
}
\arguments{
\item{grp}{an IFA group}

\item{excol}{vector of column names to omit}
}
\description{
Omit the given items
}

