% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{omitMostMissing}
\alias{omitMostMissing}
\title{Omit items with the most missing data}
\usage{
omitMostMissing(grp, omit)
}
\arguments{
\item{grp}{an IFA group}

\item{omit}{the maximum number of items to omit}
}
\description{
Items with no missing data are never omitted, regardless of the
number of items requested.
}

