% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{ordinal.gamma}
\alias{ordinal.gamma}
\title{Compute the ordinal gamma association statistic}
\usage{
ordinal.gamma(mat)
}
\arguments{
\item{mat}{a cross tabulation matrix}
}
\description{
Compute the ordinal gamma association statistic
}
\examples{
# Example data from Agresti (1990, p. 21)
jobsat <- matrix(c(20,22,13,7,24,38,28,18,80,104,81,54,82,125,113,92), nrow=4, ncol=4)
ordinal.gamma(jobsat)
}
\references{
Agresti, A. (1990). Categorical data analysis. New York: Wiley.
}
