% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flexmirt.R
\name{read.flexmirt}
\alias{read.flexmirt}
\title{Read a flexMIRT PRM file}
\usage{
read.flexmirt(fname)
}
\arguments{
\item{fname}{file name}
}
\value{
a list of groups each with item parameters and the latent distribution
}
\description{
Load the item parameters from a flexMIRT PRM file.
}

