% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.1dim}
\alias{Class rpf.1dim}
\alias{rpf.1dim-class}
\title{The base class for 1 dimensional response probability functions.}
\description{
The base class for 1 dimensional response probability functions.
}
