% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.1dim.drm}
\alias{Class rpf.1dim.drm}
\alias{rpf.1dim.drm-class}
\title{Unidimensional dichotomous item models (1PL, 2PL, and 3PL).}
\description{
Unidimensional dichotomous item models (1PL, 2PL, and 3PL).
}
