% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.1dim.gpcmp}
\alias{Class rpf.1dim.gpcmp}
\alias{rpf.1dim.gpcmp-class}
\title{Unidimensional generalized partial credit monotonic polynomial.}
\description{
Unidimensional generalized partial credit monotonic polynomial.
}
