% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.1dim.grm}
\alias{Class rpf.1dim.grm}
\alias{rpf.1dim.grm-class}
\title{The unidimensional graded response item model.}
\description{
The unidimensional graded response item model.
}
