% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.1dim.grmp}
\alias{Class rpf.1dim.grmp}
\alias{rpf.1dim.grmp-class}
\title{Unidimensional graded response monotonic polynomial.}
\description{
Unidimensional graded response monotonic polynomial.
}
