% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.1dim.lmp}
\alias{Class rpf.1dim.lmp}
\alias{rpf.1dim.lmp-class}
\title{Unidimensional logistic function of a monotonic polynomial.}
\description{
Unidimensional logistic function of a monotonic polynomial.
}
