% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.base}
\alias{Class rpf.base}
\alias{rpf.base-class}
\alias{$,rpf.base-method}
\alias{$<-,rpf.base-method}
\title{The base class for response probability functions.}
\description{
Item specifications should not be modified after creation.
}
