% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rpf.dTheta}
\alias{rpf.dTheta}
\alias{_rpf_dTheta}
\alias{rpf.dTheta,rpf.base,numeric,numeric,numeric-method}
\alias{rpf.dTheta,rpf.base,numeric,matrix,numeric-method}
\title{Item derivatives with respect to the location in the latent space}
\usage{
rpf.dTheta(m, param, where, dir)
}
\arguments{
\item{m}{item model}

\item{param}{item parameters}

\item{where}{location in the latent distribution}

\item{dir}{if more than 1 factor, a basis vector}
}
\description{
Evaluate the partial derivatives of the response probability with
respect to ability. See \link{rpf.info} for an application.
}
