\docType{methods}
\name{rpf.getLocation}
\alias{rpf.getLocation}
\alias{rpf.getLocation,rpf.1dim.drm,numeric-method}
\alias{rpf.getLocation,rpf.1dim.gpcm,numeric-method}
\alias{rpf.getLocation,rpf.mdim.drm,numeric-method}
\alias{rpf.getLocation,rpf.mdim.gpcm,numeric-method}
\alias{rpf.getLocation,rpf.mdim.mcm,numeric-method}
\alias{rpf.getLocation,rpf.mdim.nrm,numeric-method}
\title{Extract location related item parameters}
\arguments{
  \item{m}{an item model}

  \item{param}{item parameters}
}
\value{
  location related item parameters
}
\description{
  Extract location related item parameters
}
\seealso{
  \code{\link{rpf.setLocation}}
}

