\name{rpf.id_of}
\alias{rpf.id_of}
\title{Convert an IRT item model name to an ID}
\usage{
  rpf.id_of(name)
}
\arguments{
  \item{name}{name of the item model (string)}
}
\value{
  the integer ID assigned to the given model
}
\description{
  drm1 is the standard 3PL. drm is the multidimensional
  version of the 3PL. gpcm1 is the Generalized Partial
  Credit Model.
}

