% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.mdim}
\alias{Class rpf.mdim}
\alias{rpf.mdim-class}
\title{The base class for multi-dimensional response probability functions.}
\description{
The base class for multi-dimensional response probability functions.
}
