% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.mdim.drm}
\alias{Class rpf.mdim.drm}
\alias{rpf.mdim.drm-class}
\title{Multidimensional dichotomous item models (M1PL, M2PL, and M3PL).}
\description{
Multidimensional dichotomous item models (M1PL, M2PL, and M3PL).
}
