% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.mdim.grm}
\alias{Class rpf.mdim.grm}
\alias{rpf.mdim.grm-class}
\title{The multidimensional graded response item model.}
\description{
The multidimensional graded response item model.
}
