% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.mdim.mcm}
\alias{Class rpf.mdim.mcm}
\alias{rpf.mdim.mcm-class}
\title{The multiple-choice response item model (both unidimensional and
multidimensional models have the same parameterization).}
\description{
The multiple-choice response item model (both unidimensional and
multidimensional models have the same parameterization).
}
