% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{rpf.mean.info}
\alias{rpf.mean.info}
\title{Find the point where an item provides mean maximum information}
\usage{
rpf.mean.info(spec, param, grain = 0.1)
}
\arguments{
\item{spec}{list of item specs}

\item{param}{list or matrix of item parameters}

\item{grain}{the step size for numerical integration (optional)}
}
\description{
\lifecycle{experimental}
This is a point estimate of the mean difficulty of items that do
not offer easily interpretable parameters such as the Generalized
PCM. Since the information curve may not be unimodal, this
function integrates across the latent space.
}
