% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rpf.modify}
\alias{rpf.modify}
\alias{rpf.modify,rpf.mdim.drm,numeric-method}
\alias{rpf.modify,rpf.mdim.graded,numeric-method}
\alias{rpf.modify,rpf.mdim.nrm,numeric-method}
\title{Create a similar item specification with the given number of factors}
\usage{
rpf.modify(m, factors)
}
\arguments{
\item{m}{item model}

\item{factors}{the number of factors/dimensions}
}
\description{
Create a similar item specification with the given number of factors
}
\examples{
s1 <- rpf.grm(factors=3)
rpf.rparam(s1)
s2 <- rpf.modify(s1, 1)
rpf.rparam(s2)
}
