% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rpf.numParam}
\alias{rpf.numParam}
\alias{rpf.numParam,rpf.base-method}
\alias{rpf_numParam_wrapper}
\title{Length of the item parameter vector}
\usage{
rpf.numParam(m)
}
\arguments{
\item{m}{item model}
}
\description{
Length of the item parameter vector
}
\examples{
rpf.numParam(rpf.grm(outcomes=3))
rpf.numParam(rpf.nrm(outcomes=3))
}
