% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rpf.numSpec}
\alias{rpf.numSpec}
\alias{rpf.numSpec,rpf.base-method}
\alias{rpf_numSpec_wrapper}
\title{Length of the item model vector}
\usage{
rpf.numSpec(m)
}
\arguments{
\item{m}{item model}
}
\description{
Length of the item model vector
}
\examples{
rpf.numSpec(rpf.grm(outcomes=3))
rpf.numSpec(rpf.nrm(outcomes=3))
}
