% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rpf.paramInfo}
\alias{rpf.paramInfo}
\alias{rpf.paramInfo,rpf.base-method}
\alias{rpf_paramInfo_wrapper}
\title{Retrieve a description of the given parameter}
\usage{
rpf.paramInfo(m, num = NULL)
}
\arguments{
\item{m}{item model}

\item{num}{vector of parameters (defaults to all)}
}
\value{
a list containing the type, upper bound, and lower bound
}
\description{
Retrieve a description of the given parameter
}
\examples{
rpf.paramInfo(rpf.drm())
}
