% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{rpf.rparam}
\alias{rpf.rparam}
\alias{rpf.rparam,rpf.1dim.drm-method}
\alias{rpf.rparam,rpf.mdim.drm-method}
\alias{rpf.rparam,rpf.1dim.graded-method}
\alias{rpf.rparam,rpf.mdim.graded-method}
\alias{rpf.rparam,rpf.mdim.nrm-method}
\alias{rpf.rparam,rpf.mdim.mcm-method}
\alias{rpf.rparam,rpf.1dim.lmp-method}
\alias{rpf.rparam,rpf.1dim.grmp-method}
\alias{rpf.rparam,rpf.1dim.gpcmp-method}
\title{Generates item parameters}
\usage{
rpf.rparam(m, version = 2L)
}
\arguments{
\item{m}{an item model}

\item{version}{the version of random parameters}
}
\value{
item parameters
}
\description{
This function generates random item parameters. The version
argument is available if you are writing a test that depends on
reproducable random parameters (using \code{set.seed}).
}
\examples{
i1 <- rpf.drm()
rpf.rparam(i1)
}
