\docType{data}
\name{science}
\alias{science}
\alias{science.items}
\alias{science.people}
\title{Liking for Science dataset}
\description{
  These data are from Wright & Masters (1982, p. 18).
}
\details{
  All items were fit to a 3 category Partial Credit Model
  (PCM) with standard Gauss-Hermite quadrature of 49 points
  using OpenMx, -2loglikelihood = 2733.3. Person scores
  were obtained using EAP with a 98 point flat quadrature.
}
\examples{
data(science.items)
data(science.people)
}
\references{
  Wright, B. D. & Masters, G. N. (1982). \emph{Rating Scale
  Analysis.} Chicago: Mesa Press.
}
\keyword{datasets}

