% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ssEAP}
\alias{ssEAP}
\title{Internal sum-score EAP engine}
\usage{
ssEAP(grp, qwidth, qpoints, mask, twotier = FALSE, debug = FALSE)
}
\arguments{
\item{grp}{grp}

\item{qwidth}{qwidth}

\item{qpoints}{qpoints}

\item{mask}{mask}

\item{twotier}{twotier}

\item{debug}{debug}
}
\description{
Do not call this directly. For developers only
}

