% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openmx.R
\name{stripData}
\alias{stripData}
\title{Strip data and scores from an IFA group}
\usage{
stripData(grp)
}
\arguments{
\item{grp}{an IFA group}
}
\description{
In addition, the weightColumn is reset to NULL.
}

