% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sumScoreEAPTest}
\alias{sumScoreEAPTest}
\title{Conduct the sum-score EAP distribution test}
\usage{
sumScoreEAPTest(grp, ..., .twotier = TRUE)
}
\arguments{
\item{grp}{a list with spec, param, mean, and cov}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{.twotier}{whether to enable the two-tier optimization}
}
\description{
Conduct the sum-score EAP distribution test
}
\references{
Li, Z., & Cai, L. (2012, July). Summed score likelihood based indices for testing
latent variable distribution fit in Item Response Theory. Paper presented at
the annual International Meeting of the Psychometric Society, Lincoln,
NE. Retrieved from http://www.cse.ucla.edu/downloads/files/SD2-final-4.pdf
}

