% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{tabulateRows}
\alias{tabulateRows}
\title{Tabulate data.frame rows}
\usage{
tabulateRows(observed)
}
\arguments{
\item{observed}{a sorted data.frame holding ordered factors in every column}
}
\description{
Like \code{tabulate} but entire rows are the unit of tabulation.
The data.frame is not sorted, but must be sorted already.
}
\examples{
df <- as.data.frame(matrix(c(sample.int(2, 30, replace=TRUE)), 10, 3))
df <- df[orderCompletely(df),]
tabulateRows(df)
}
\seealso{
\code{\link{orderCompletely}}
}
