% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toFactorLoading}
\alias{toFactorLoading}
\title{Convert response function slopes to factor loadings}
\usage{
toFactorLoading(slope, ogive = rpf.ogive)
}
\arguments{
\item{slope}{a matrix with items in the columns and slopes in the rows}

\item{ogive}{the ogive constant (default rpf.ogive)}
}
\value{
a factor loading matrix with items in the rows and factors in the columns
}
\description{
Convert response function slopes to factor loadings
}

