% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{toFactorThreshold}
\alias{toFactorThreshold}
\title{Convert response function intercepts to factor thresholds}
\usage{
toFactorThreshold(intercept, slope, ogive = rpf.ogive)
}
\arguments{
\item{intercept}{a matrix with items in the columns and intercepts in the rows}

\item{slope}{a matrix with items in the columns and slopes in the rows}

\item{ogive}{the ogive constant (default \link{rpf.ogive})}
}
\value{
a factor threshold matrix with items in the columns and factor thresholds in the rows
}
\description{
Convert response function intercepts to factor thresholds
}
\seealso{
Other factor model equivalence: 
\code{\link{fromFactorLoading}()},
\code{\link{fromFactorThreshold}()},
\code{\link{toFactorLoading}()}
}
\concept{factor model equivalence}
