% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.IFAgroup}
\alias{as.IFAgroup}
\title{Convert an OpenMx MxModel object into an IFA group}
\usage{
as.IFAgroup(mxModel, data = NULL, container = NULL, ...,
  minItemsPerScore = NULL)
}
\arguments{
\item{mxModel}{MxModel object}

\item{data}{observed data (otherwise the data will be taken from the mxModel)}

\item{container}{an MxModel in which to search for the latent distribution matrices}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{minItemsPerScore}{minimum number of items required to compute a score (also see description)}
}
\value{
a groups with item parameters and latent distribution
}
\description{
When \dQuote{minItemsPerScore} is passed, EAP scores will be computed
from the data and stored. Scores are required for some diagnostic
tests. See discussion of \dQuote{minItemsPerScore} in \link{EAPscores}.
}
\seealso{
\href{https://github.com/jpritikin/ifaTools}{ifaTools}
}

