% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.IFAgroup}
\alias{as.IFAgroup}
\title{Convert an OpenMx MxModel object into an IFA group}
\usage{
as.IFAgroup(mxModel, data = NULL)
}
\arguments{
\item{mxModel}{MxModel object}

\item{data}{observed data (otherwise the data will be taken from the mxModel)}
}
\value{
a groups with item parameters and latent distribution
}
\description{
Convert an OpenMx MxModel object into an IFA group
}

