% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compressDataFrame}
\alias{compressDataFrame}
\title{Compress a data frame into unique rows and frequencies}
\usage{
compressDataFrame(tabdata, freqColName = "freq")
}
\arguments{
\item{tabdata}{An object of class \code{data.frame}}

\item{freqColName}{Column name to contain the frequencies}
}
\value{
Returns a compressed data frame
}
\description{
Compress a data frame into unique rows and frequency counts.
}
\examples{
df <- as.data.frame(matrix(c(sample.int(2, 30, replace=TRUE)), 10, 3))
compressDataFrame(df)
}

