% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crosstabTest}
\alias{crosstabTest}
\title{Monte-Carlo test for cross-tabulation tables}
\usage{
crosstabTest(ob, ex, trials)
}
\arguments{
\item{ob}{observed table}

\item{ex}{expected table}

\item{trials}{number of Monte-Carlo trials}
}
\description{
This is for developers.
}

