% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expandDataFrame}
\alias{expandDataFrame}
\title{Expand summary table of patterns and frequencies}
\usage{
expandDataFrame(tabdata, freqName = NULL)
}
\arguments{
\item{tabdata}{An object of class \code{data.frame} with the unique response patterns and the number of frequencies}

\item{freqName}{Column name containing the frequencies}
}
\value{
Returns a data frame with all the response patterns
}
\description{
Expand a summary table of unique response patterns to a full sized
data-set.
}
\author{
Based on code by Phil Chalmers \email{rphilip.chalmers@gmail.com}
}

