% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fromFactorLoading}
\alias{fromFactorLoading}
\title{Convert factor loadings to response function slopes}
\usage{
fromFactorLoading(loading, ogive = rpf.ogive)
}
\arguments{
\item{loading}{a matrix with items in the rows and factors in the columns}

\item{ogive}{the ogive constant (default \link{rpf.ogive})}
}
\value{
a slope matrix with items in the columns and factors in the rows
}
\description{
Convert factor loadings to response function slopes
}
\seealso{
Other factor model equivalence: 
\code{\link{fromFactorThreshold}()},
\code{\link{toFactorLoading}()},
\code{\link{toFactorThreshold}()}
}
\concept{factor model equivalence}
