% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fromFactorThreshold}
\alias{fromFactorThreshold}
\title{Convert factor thresholds to response function intercepts}
\usage{
fromFactorThreshold(threshold, loading, ogive = rpf.ogive)
}
\arguments{
\item{threshold}{a matrix with items in the columns and thresholds in the rows}

\item{loading}{a matrix with items in the rows and factors in the columns}

\item{ogive}{the ogive constant (default rpf.ogive)}
}
\value{
an item intercept matrix with items in the columns and intercepts in the rows
}
\description{
Convert factor thresholds to response function intercepts
}

