\name{ordinal.gamma}
\alias{ordinal.gamma}
\title{Compute the ordinal gamma association statistic}
\usage{
  ordinal.gamma(mat)
}
\arguments{
  \item{mat}{a cross tabulation matrix}
}
\description{
  Compute the ordinal gamma association statistic
}
\references{
  Agresti, A. (1990). Categorical data analysis. New York:
  Wiley.
}

