% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Class rpf.base}
\alias{$,rpf.base-method}
\alias{$<-,rpf.base-method}
\alias{Class rpf.base}
\alias{rpf.base-class}
\title{The base class for response probability functions.}
\description{
Item specifications should not be modified after creation.
}

