\name{rpf.dTheta}
\alias{rpf.dTheta}
\alias{rpf.dTheta,rpf.base,numeric,matrix,numeric-method}
\alias{rpf.dTheta,rpf.base,numeric,numeric,numeric-method}
\alias{rpf_dTheta_wrapper}
\title{Item derivatives with respect to ability}
\arguments{
  \item{m}{item model}

  \item{param}{item parameters}

  \item{where}{location in the latent distribution}

  \item{dir}{if more than 1 factor, a basis vector]}
}
\description{
  Evaluate the partial derivatives of the response
  probability with respect to ability.
}

