\docType{methods}
\name{rpf.getLocation}
\alias{rpf.getLocation}
\alias{rpf.getLocation,rpf.drm,numeric-method}
\alias{rpf.getLocation,rpf.gpcm,numeric-method}
\title{Extract location related item parameters}
\arguments{
  \item{m}{an item model}

  \item{param}{item parameters}
}
\value{
  location related item parameters
}
\description{
  Extract location related item parameters
}
\seealso{
  \code{\link{rpf.setLocation}}
}

