\name{rpf.gradient}
\alias{rpf.gradient}
\alias{rpf.gradient,rpf.base,numeric,integer,numeric,numeric-method}
\alias{rpf_gradient_wrapper}
\title{Item parameter gradients}
\description{
  Evaluate the partial derivatives of the log likelihood of
  the item model with respect to each parameter at
  \code{where} with \code{weight}.
}

