% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rpf.id_of}
\alias{rpf.id_of}
\title{Convert an rpf item model name to an ID}
\usage{
rpf.id_of(name)
}
\arguments{
\item{name}{name of the item model (string)}
}
\value{
the integer ID assigned to the given model
}
\description{
This is an internal function and should not be used.
}
