% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcm.R
\name{rpf.mcm}
\alias{rpf.mcm}
\title{Create a multiple-choice response model}
\usage{
rpf.mcm(outcomes = 2, numChoices = 5, factors = 1)
}
\arguments{
\item{outcomes}{the number of possible outcomes}

\item{numChoices}{the number of choices available}

\item{factors}{the number of factors}
}
\value{
an item model
}
\description{
\lifecycle{experimental}
}
\details{
This function instantiates a multiple-choice response
model.
}
\seealso{
Other response model: 
\code{\link{rpf.drm}()},
\code{\link{rpf.gpcmp}()},
\code{\link{rpf.grmp}()},
\code{\link{rpf.grm}()},
\code{\link{rpf.lmp}()},
\code{\link{rpf.nrm}()}
}
\author{
Jonathan Weeks <weeksjp@gmail.com>
}
\concept{response model}
