% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rpf.mcm}
\alias{rpf.mcm}
\title{Create a multiple-choice response model}
\usage{
rpf.mcm(outcomes = 2, numChoices = 5, factors = 1)
}
\arguments{
\item{outcomes}{the number of possible outcomes}

\item{numChoices}{the number of choices available}

\item{factors}{the number of factors}
}
\value{
an item model
}
\description{
WARNING: This model is mostly not implemented.
}
\details{
This function instantiates a multiple-choice response
model.
}
\author{
Jonathan Weeks <weeksjp@gmail.com>
}

