% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Class rpf.mdim}
\alias{Class rpf.mdim}
\alias{rpf.mdim-class}
\title{The base class for multi-dimensional response probability functions.}
\description{
The base class for multi-dimensional response probability functions.
}

