\docType{class}
\name{Class rpf.mdim.mcm}
\alias{Class rpf.mdim.mcm}
\alias{rpf.mdim.mcm-class}
\title{The multiple-choice response item model (both unidimensional and
multidimensional models have the same parameterization).}
\description{
The multiple-choice response item model (both
unidimensional and multidimensional models have the same
parameterization).
}

