% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Class rpf.mdim.nrm}
\alias{Class rpf.mdim.nrm}
\alias{rpf.mdim.nrm-class}
\title{The nominal response item model (both unidimensional and
multidimensional models have the same parameterization).}
\description{
The nominal response item model (both unidimensional and
multidimensional models have the same parameterization).
}
