% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rpf.numSpec}
\alias{rpf.numSpec}
\alias{rpf.numSpec,rpf.base-method}
\alias{rpf_numSpec_wrapper}
\title{Length of the item model vector}
\usage{
rpf.numSpec(m)
}
\arguments{
\item{m}{item model}
}
\description{
Length of the item model vector
}

