\docType{data}
\name{rpf.ogive}
\alias{rpf.ogive}
\title{The ogive constant}
\format{num 1.7}
\usage{
  rpf.ogive
}
\description{
  The ogive constant can be multiplied by the
  discrimination parameter to obtain a response curve very
  similar to the Normal cumulative distribution function
  (Haley, 1952; Molenaar, 1974). In recent years, the
  logistic has grown in favor, and therefore, this package
  does not offer any special support for this
  transformation (Baker & Kim, 2004, pp. 14-18).
}
\references{
  Baker & Kim (2004). \emph{Item Response Theory: Parameter
  Estimation Techniques.} Marcel Dekker, Inc.

  Haley, D. C. (1952). \emph{Estimation of the dosage
  mortality relationship when the dose is subject to error}
  (Technical Report No. 15). Stanford University Applied
  Mathematics and Statistics Laboratory, Stanford, CA.

  Molenaar, W. (1974). De logistische en de normale kromme
  [The logistic and the normal curve]. \emph{Nederlands
  Tijdschrift voor de Psychologie} 29, 415-420.
}
\keyword{datasets}

