\docType{data}
\name{rpf.ogive}
\alias{rpf.ogive}
\title{The ogive constant}
\format{num 1.7}
\usage{
  rpf.ogive
}
\description{
  Models built on the logistic function take an argument
  \code{D} where you can pass in the ogive constant to
  obtain a response curve very similar to the Normal
  cumulative distribution function. In recent years, the
  logistic has grown in favor, and therefore, \code{D}
  defaults to 1 in this package (Baker & Kim, 2004, pp.
  14-18).
}
\references{
  Baker & Kim (2004). Item Response Theory: Parameter
  Estimation Techniques. Marcel Dekker, Inc.
}
\keyword{datasets}

