\name{rpf.rescale}
\alias{rpf.rescale}
\alias{rpf.rescale,rpf.base,numeric,numeric,matrix-method}
\alias{rpf_rescale_wrapper}
\title{Rescale item parameters}
\arguments{
  \item{m}{item model}

  \item{param}{item parameters}

  \item{mean}{vector of means}

  \item{cov}{covariance matrix}
}
\description{
  Adjust item parameters for changes in mean and covariance
  of the latent distribution.
}

