\docType{methods}
\name{rpf.setLocation}
\alias{rpf.setLocation}
\alias{rpf.setLocation,rpf.drm,numeric,numeric-method}
\alias{rpf.setLocation,rpf.gpcm,numeric,numeric-method}
\title{Set location related item parameters}
\arguments{
  \item{m}{an item model}

  \item{param}{item parameters}

  \item{loc}{location parameters, typically from
  \code{\link{rpf.getLocation}}}
}
\value{
  updated item parameters
}
\description{
  Set location related item parameters
}
\seealso{
  \code{\link{rpf.getLocation}}
}

