% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/science.R
\docType{data}
\name{science}
\alias{science}
\alias{sfif}
\alias{sfpf}
\alias{sfsf}
\alias{sfxf}
\title{Liking for Science dataset}
\description{
These data are from Wright & Masters (1982, p. 18).
}
\details{
All items were fit to a 3 category Partial Credit Model (PCM)
using Ministep 3.75.0.
}
\examples{
data(science)
}
\references{
Wright, B. D. & Masters, G. N. (1982). \emph{Rating Scale
Analysis.} Chicago: Mesa Press.
}
\keyword{data}
