% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tabulateRows}
\alias{tabulateRows}
\title{Tabulate data.frame rows}
\usage{
tabulateRows(observed)
}
\arguments{
\item{observed}{a sorted data.frame holding ordered factors in every column}
}
\description{
Like \code{tabulate} but entire rows are the unit of tabulation.
The data.frame is not sorted, but must be sorted already.
}
\seealso{
\code{\link{orderCompletely}}
}

